requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests
python-telegram-bot==13.11.0
SQLAlchemy==1.4.32
psycopg2-binary
aiofiles
feedparser
faker
hachoir
pykeyboard
pynewtonmath
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans==4.0.0rc1
google-trans-new
jikanpy
speedtest-cli
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==1.4.9

git+https://github.com/Sorrybabe/downloader
asyncio
cloudscraper
dateparser
pymongo
dnspython
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
pytimeparse
PyYAML>=5.3.1
